+++ decompiler/IR2/FormStack.cpp
@@ -1,3 +1,4 @@
+#include <algorithm>
 #include "FormStack.h"
 #include "Form.h"